Properly implemented the game timer.
But unfortunately this does not solve all of our timing problems because the loop won't be called
EXACTLY the same time each time. Therefore we need to modify all movements to incorporate the
current tick time. So that when the tick time it a bit higher everything thing will move a bit
further.